refactor java version management
