+++ src/commands/java/deactivate.rs
