+++ src/commands/java/mod.rs
@@ -1,0 +1,8 @@
+pub mod deactivate;
+pub mod install;
+pub mod usev;
+pub mod lsrem;
+pub mod current;
+pub mod ls;
+pub mod uninstall;
+pub mod clean;
