+++ src/commands/java/uninstall.rs
