+++ src/commands/mod.rs
@@ -1,8 +1,1 @@
-pub mod clean;
-pub mod current;
-pub mod deactivate;
-pub mod install;
-pub mod ls;
-pub mod lsrem;
-pub mod uninstall;
-pub mod usev;
+pub mod java;
