+++ src/constants/mod.rs
@@ -1,1 +1,1 @@
-pub mod versions;
+pub mod java_versions;
